Сделала адаптивными модалки корзины и оформления заказа. Сделала кнопку Назад в карточке товара. Полностью сделана админка с авторизацией. Вывод состава заказа, сделаны дополнительные кнопки, на русском языке, стандартизирован вид на страницах админки.

+++ controllers/AdminController.php
@@ -5,9 +5,12 @@ namespace app\controllers;
 use Yii;
 use app\models\Order;
 use yii\data\ActiveDataProvider;
+use yii\helpers\Url;
 use yii\web\Controller;
 use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
+use yii\filters\AccessControl;
+use app\models\LoginForm;
 
 /** AdminController implements the CRUD actions for Order model. */
 class AdminController extends Controller
@@ -16,25 +19,72 @@ class AdminController extends Controller
     public function behaviors()
     {
         return [
+            'access' => [
+                'class' => AccessControl::class,
+                'only' => ['logout'],
+                'rules' => [
+                    [
+                        'actions' => ['logout'],
+                        'allow' => true,
+                        'roles' => ['@'],
+                    ],
+                ],
+            ],
             'verbs' => [
-                'class' => VerbFilter::className(),
+                'class' => VerbFilter::class,
                 'actions' => [
                     'delete' => ['POST'],
+                    'logout' => ['POST', 'GET'],
                 ],
             ],
         ];
     }
 
+    public function actionLogin()
+    {
+        $this->layout = 'admin-layout';
+        if (!Yii::$app->user->isGuest) {
+            return $this->redirect([Url::to('/admin/')]);
+
+            /*$dataProvider = new ActiveDataProvider([
+                'query' => Order::find(),
+            ]);
+            return $this->render('index', compact('dataProvider'));*/
+        }
+
+        $model = new LoginForm();
+        if ($model->load(Yii::$app->request->post()) && $model->login()) {
+            $dataProvider = new ActiveDataProvider([
+                'query' => Order::find(),
+            ]);
+            return $this->redirect([Url::to('/admin/')]);
+//            return $this->render('index', compact('dataProvider'));
+        }
+
+        $model->password = '';
+        return $this->render('login', [
+            'model' => $model,
+        ]);
+    }
+
+    public function actionLogout()
+    {
+        Yii::$app->user->logout();
+        return $this->goHome();
+    }
+
     public function actionIndex()
     {
-        $dataProvider = new ActiveDataProvider([
-            'query' => Order::find(),
-        ]);
-        $this->layout = 'admin-layout';
+        if (!Yii::$app->user->isGuest) {
+            $dataProvider = new ActiveDataProvider([
+                'query' => Order::find(),
+            ]);
+            $this->layout = 'admin-layout';
 
-        return $this->render('index', [
-            'dataProvider' => $dataProvider,
-        ]);
+            return $this->render('index', compact('dataProvider'));
+        } else {
+            return $this->goHome();
+        }
     }
 
     /**
@@ -45,23 +95,14 @@ class AdminController extends Controller
      */
     public function actionView($id)
     {
-        $this->layout = 'admin-layout';
-        return $this->render('view', [
-            'model' => $this->findModel($id),
-        ]);
-    }
-
-    public function actionCreate()
-    {
-        $model = new Order();
-
-        if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'id' => $model->id]);
+        if (!Yii::$app->user->isGuest) {
+            $this->layout = 'admin-layout';
+            return $this->render('view', [
+                'model' => $this->findModel($id),
+            ]);
+        } else {
+            return $this->goHome();
         }
-
-        return $this->render('create', [
-            'model' => $model,
-        ]);
     }
 
     /**
@@ -73,15 +114,19 @@ class AdminController extends Controller
      */
     public function actionUpdate($id)
     {
-        $model = $this->findModel($id);
+        if (!Yii::$app->user->isGuest) {
+            $model = $this->findModel($id);
+            $this->layout = 'admin-layout';
 
-        if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'id' => $model->id]);
+            if ($model->load(Yii::$app->request->post()) && $model->save()) {
+                return $this->redirect(['view', 'id' => $model->id]);
+            }
+            return $this->render('update', [
+                'model' => $model,
+            ]);
+        } else {
+            return $this->goHome();
         }
-        $this->layout = 'admin-layout';
-        return $this->render('update', [
-            'model' => $model,
-        ]);
     }
 
     /**
@@ -93,9 +138,12 @@ class AdminController extends Controller
      */
     public function actionDelete($id)
     {
-        $this->findModel($id)->delete();
-
-        return $this->redirect(['index']);
+        if (!Yii::$app->user->isGuest) {
+            $this->findModel($id)->delete();
+            return $this->redirect(['index']);
+        } else {
+            return $this->goHome();
+        }
     }
 
     /**
@@ -110,7 +158,6 @@ class AdminController extends Controller
         if (($model = Order::findOne($id)) !== null) {
             return $model;
         }
-
         throw new NotFoundHttpException('The requested page does not exist.');
     }
 }
